Moved DeliminatorInfo class
