+++ Sharpdown/MarkdownElement/InlineElement/InlineParserObjects/DeliminatorType.cs
@@ -1,0 +1,25 @@
+namespace Sharpdown.MarkdownElement.InlineElement.InlineParserObjects
+{
+    internal enum DeliminatorType
+    {
+        /// <summary>
+        /// Open Link ([).
+        /// </summary>
+        OpenLink,
+
+        /// <summary>
+        /// Open Image (![).
+        /// </summary>
+        OpenImage,
+
+        /// <summary>
+        /// Star (*).
+        /// </summary>
+        Star,
+
+        /// <summary>
+        /// UnderBar (_).
+        /// </summary>
+        UnderBar,
+    }
+}
